+++ lib/surgex/parser/parsers/boolean_parser.ex
@@ -3,8 +3,11 @@ defmodule Surgex.Parser.BooleanParser do
 
   @spec call(term()) :: {:ok, boolean() | nil} | {:error, :invalid_boolean}
   def call(nil), do: {:ok, nil}
+  def call(""), do: {:ok, nil}
   def call("0"), do: {:ok, false}
   def call("1"), do: {:ok, true}
+  def call("false"), do: {:ok, false}
+  def call("true"), do: {:ok, true}
   def call(false), do: {:ok, false}
   def call(true), do: {:ok, true}
   def call(_input), do: {:error, :invalid_boolean}