+++ lib/surgex/parser/parsers/geobox_parser.ex
@@ -7,6 +7,7 @@ defmodule Surgex.Parser.GeoboxParser do
 
   @spec call(term()) :: {:ok, Geobox.t() | nil} | {:error, errors}
   def call(nil), do: {:ok, nil}
+  def call(""), do: {:ok, nil}
 
   # credo:disable-for-next-line Credo.Check.Refactor.ABCSize
   def call(input) when is_binary(input) do