+++ lib/surgex/parser/parsers/resource_id_parser.ex
@@ -6,6 +6,8 @@ defmodule Surgex.Parser.ResourceIdParser do
   @spec call(term()) :: {:ok, integer | nil} | {:error, Keyword.t()}
   def call(nil), do: {:ok, nil}
 
+  def call(%{id: ""}), do: {:error, [required: "id"]}
+
   def call(%{id: id_string}) when is_binary(id_string) do
     with {:error, reason} <- IdParser.call(id_string) do
       {:error, [{reason, "id"}]}