+++ lib/surgex/parser/parsers/slug_parser.ex
@@ -3,6 +3,7 @@ defmodule Surgex.Parser.SlugParser do
 
   @spec call(term()) :: {:ok, String.t() | nil} | {:error, :invalid_slug}
   def call(nil), do: {:ok, nil}
+  def call(""), do: {:ok, nil}
 
   def call(input) when is_binary(input) do
     if String.match?(input, ~r/^[a-zA-Z0-9\-]+$/) do