+++ lib/surgex/parser/parsers/sort_parser.ex
@@ -10,6 +10,7 @@ defmodule Surgex.Parser.SortParser do
   @doc false
   @spec call(term(), [atom]) :: {:ok, {:asc | :desc, atom} | nil} | {:error, :invalid_sort_column}
   def call(nil, _allowed_columns), do: {:ok, nil}
+  def call("", _allowed_columns), do: {:ok, nil}
 
   def call(input, allowed_columns) when is_binary(input) do
     case input do