+++ test/surgex/parser/parsers/boolean_parser_test.exs
@@ -6,9 +6,15 @@ defmodule Surgex.Parser.BooleanParserTest do
     assert BooleanParser.call(nil) == {:ok, nil}
   end
 
+  test "empty string" do
+    assert BooleanParser.call(nil) == {:ok, nil}
+  end
+
   test "valid input" do
     assert BooleanParser.call("0") == {:ok, false}
     assert BooleanParser.call("1") == {:ok, true}
+    assert BooleanParser.call("false") == {:ok, false}
+    assert BooleanParser.call("true") == {:ok, true}
     assert BooleanParser.call(false) == {:ok, false}
     assert BooleanParser.call(true) == {:ok, true}
   end